test: hoist tests/fast conftest to tests/, privatize e2e dumper conftest
